organize main component
