+++ frontend/src/App.tsx
@@ -1,44 +1,31 @@
-import "bootstrap/dist/css/bootstrap.min.css";
-import React, { useState, useReducer } from "react";
-import { ChakraProvider, extendTheme } from "@chakra-ui/react";
+import React from 'react';
 import {
   BrowserRouter as Router,
   Route,
   Routes as Switch,
 } from "react-router-dom";
 
-import * as Routes from "./constants/Routes";
-import { getLocalStorageObj } from "./utils/LocalStorageUtils";
-import SampleContext, {
-  DEFAULT_SAMPLE_CONTEXT,
-} from "./contexts/SampleContext";
-import SampleContextDispatcherContext from "./contexts/SampleContextDispatcherContext";
-import sampleContextReducer from "./reducers/SampleContextReducer";
+import { ApolloClient, ApolloProvider, InMemoryCache } from "@apollo/client";
+import { createUploadLink } from "apollo-upload-client";
+import { setContext } from "@apollo/client/link/context";
 
+import "bootstrap/dist/css/bootstrap.min.css";
+import { ChakraProvider, extendTheme } from "@chakra-ui/react";
 import colors from "./theme/colors";
 import modalTheme from "./theme/modals";
 import buttonTheme from "./theme/buttons";
 import tabsTheme from "./theme/tabs";
 import { inputTheme, textareaTheme } from "./theme/inputs";
 
-import HomePage from "./components/pages/home/HomePage";
-import AnnouncementsPage from "./components/pages/announcements/AnnouncementsPage";
-import TasksPage from "./components/pages/tasks/TasksPage";
-import SchedulePage from "./components/pages/schedule/SchedulePage";
-import ResidentsPage from "./components/pages/residents/ResidentsPage";
-import InsightsPage from "./components/pages/insights/InsightsPage";
-import NotFoundPage from "./components/pages/NotFoundPage";
+import HomePage from "./components/sections/home/HomePage";
+import SchedulePage from "./components/sections/schedule/SchedulePage";
+import AnnouncementsPage from "./components/sections/announcements/AnnouncementsPage";
+import ParticipantsPage from "./components/sections/participants/ParticipantsPage";
+import TasksPage from "./components/sections/tasks/TasksPage";
+import NotFoundPage from "./components/common/NotFoundPage";
+import * as Routes from "./constants/routes";
 
 const App = (): React.ReactElement => {
-
-  // Some sort of global state. Context API replaces redux.
-  // Split related states into different contexts as necessary.
-  // Split dispatcher and state into separate contexts as necessary.
-  const [sampleContext, dispatchSampleContextUpdate] = useReducer(
-    sampleContextReducer,
-    DEFAULT_SAMPLE_CONTEXT,
-  );
-
   const theme = extendTheme({
     colors,
     components: {
@@ -50,57 +37,44 @@ const App = (): React.ReactElement => {
     },
   });
 
+  const endpoint = createUploadLink({
+    uri: `${process.env.REACT_APP_BACKEND_URL}/graphql`,
+    credentials: "include",
+  });
+
+  const header = setContext(async (_, { headers }) => {
+    // TODO: Get the authentication token from local storage
+    const token = null;
+    return {
+      headers: {
+        ...headers,
+        authorization: token ? `Bearer ${token}` : "",
+      },
+    };
+  });
+
+  const apolloClient = new ApolloClient({
+    link: header.concat(endpoint as any),
+    cache: new InMemoryCache()
+  });
+
   return (
-    <ChakraProvider theme={theme}>
-      <SampleContext.Provider value={sampleContext}>
-        <SampleContextDispatcherContext.Provider
-          value={dispatchSampleContextUpdate}
-        >
-            <Router>
-              <Switch>
-                <Route
-                  path={Routes.HOME_PAGE}
-                  element={
-                      <HomePage />
-                  }
-                />
-                <Route
-                  path={Routes.ANNOUNCEMENTS_PAGE}
-                  element={
-                      <AnnouncementsPage />
-                  }
-                />
-                <Route
-                  path={Routes.TASKS_PAGE}
-                  element={
-                      <TasksPage />
-                  }
-                />
-                <Route
-                  path={Routes.SCHEDULE_PAGE}
-                  element={
-                      <SchedulePage />
-                  }
-                />
-                <Route
-                  path={Routes.PARTICIPANTS_PAGE}
-                  element={
-                      <ResidentsPage />
-                  }
-                />
-                <Route
-                  path={Routes.INSIGHTS_PAGE}
-                  element={
-                      <InsightsPage />
-                  }
-                />
-                <Route path="*" element={<NotFoundPage />} />
-              </Switch>
-            </Router>
-        </SampleContextDispatcherContext.Provider>
-      </SampleContext.Provider>
-    </ChakraProvider>
+    <ApolloProvider client={apolloClient}>
+      <ChakraProvider theme={theme}>
+        <Router>
+          <Switch>
+            <Route path={Routes.HOME_PAGE} element={<HomePage />} />
+            <Route path={Routes.SCHEDULE_PAGE} element={<SchedulePage />} />
+            <Route path={Routes.ANNOUNCEMENTS_PAGE} element={<AnnouncementsPage />} />
+            <Route path={Routes.PARTICIPANTS_PAGE} element={<ParticipantsPage />} />
+            <Route path={Routes.TASKS_PAGE} element={<TasksPage />} />
+            <Route path="*" element={<NotFoundPage />} />
+          </Switch>
+        </Router>
+      </ChakraProvider>
+    </ApolloProvider>
   );
 };
 
 export default App;
+
