+++ frontend/src/components/sections/announcements/AnnouncementsGroups.tsx
