+++ frontend/src/components/sections/home/HomeRoomCard.tsx
