+++ frontend/src/components/sections/home/RoomGrid.tsx
