+++ frontend/src/components/sections/insights/InsightsPage.tsx
