+++ frontend/src/components/sections/participants/ResidentModal.tsx
