+++ frontend/src/components/sections/schedule/listView/ScheduleTable.tsx
